+++ examples/testing/projects.php
@@ -89,8 +89,8 @@ try {
 
     $response = $sandboxProjects->delete($projectId);
 
-    // print the response body (array)
-    var_dump(ResponseHelper::toArray($response));
+    // Print the response status code
+    var_dump($response->getStatusCode());
 } catch (Exception $e) {
     echo 'Caught exception: ',  $e->getMessage(), "\n";
 }